ocr/requirements.txt
opencv-python
pyautogui
pillow
numpy
paddleocr
shapely
pyclipper
setuptools